added lock

+++ src/client/RainLib/file_event_handler.py
@@ -10,13 +10,13 @@ class RainFileSystemEventHandler(FileSystemEventHandler):
         self.watch_dir = watch_dir
 
     def on_created(self, event):
-        if not event.is_directory:
+        if not event.is_directory and not self.rain_drive.is_locked():
             event_path = relpath(event.src_path,self.watch_dir)
             print 'create file' + event_path
             self.rain_drive.new_file(event_path)
 
     def on_deleted(self, event):
-        if not event.is_directory:
+        if not event.is_directory and not self.rain_drive.is_locked():
             event_path = relpath(event.src_path,self.watch_dir)
             print 'delete file' + event_path
             self.rain_drive.remove_file(event_path)
